feat: 拆分全局配置页面，优化结构
